+++ packages/components-react/tests/unit/helpers/index.tsx
@@ -1,0 +1,8 @@
+import { render } from '@testing-library/react';
+import { ReactElement } from 'react';
+
+export const testSnapshot = (component: ReactElement) => {
+  const { container } = render(component);
+
+  expect(container).toMatchSnapshot();
+};