+++ packages/components/CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
+#### Added
+- Link support for `Marque`
+- Sizing options `'responsive' | 'small' | 'medium'` for `Marque`
+
 ### [2.0.0-alpha.13] - 2021-01-26
 
 #### Added