+++ packages/components/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Fixed
 - Unhandled exception in `Select Wrapper` if `selected` and `disabled` attributes are set on the same option
+- Angular: typings 
 
 ### [2.0.0-alpha.13] - 2021-01-26